lowlevel: Switch declarations for go to work
Because I added libusbi.h to libusb.go, go stopped building it because of some cgo magic.

I have to admit I don't understand cgo that much. But when I shuffled around the headers it started to build again.
