+++ usb/lowlevel/libusb.go
@@ -14,7 +14,29 @@ Copyright (c) 2017 Jason T. Harris
 package lowlevel
 
 /*
+struct list_head {
+	struct list_head *prev, *next;
+};
+
 #include "./c/libusb/libusb.h"
+
+#ifdef OS_WINDOWS
+  #define usbi_mutex_t              HANDLE
+  #define usbi_tls_key_t                      DWORD
+
+  typedef struct usbi_cond {
+    struct list_head waiters;
+    struct list_head not_waiting;
+  } usbi_cond_t;
+
+  #define usbi_mutex_static_t       volatile LONG
+#else
+  #define usbi_mutex_t              pthread_mutex_t
+  #define usbi_tls_key_t                      pthread_key_t
+  #define usbi_cond_t                 pthread_cond_t
+  #define usbi_mutex_static_t         pthread_mutex_t
+#endif
+
 #include "./c/libusb/libusbi.h"
 
 // When a C struct ends with a zero-sized field, but the struct itself is not zero-sized,
